Conflicts:

KBLove.xcodeproj/xcuserdata/ming.xcuserdatad/xcschemes/xcschememanagemen
t.plist

KBLove.xcworkspace/xcuserdata/ming.xcuserdatad/UserInterfaceState.xcuser
state

Pods/Pods.xcodeproj/xcuserdata/ming.xcuserdatad/xcschemes/xcschememanage
ment.plist

+++ KBLove/Classes/Sections/Main/Controllers/ZWLGaoDeMap_TrackerReplayViewController.mm
@@ -186,12 +186,11 @@
     
     if (_isViewAllTrack) {
         [_mapView addOverlays:_allTrackPath];
-        
     } else {
         NSInteger size = _statusArray.count;
         CLLocationCoordinate2D* points = new CLLocationCoordinate2D[size];
         
-        for(int idx = size - 1; idx >= 0; idx--)
+        for(int idx = (int)size - 1; idx >= 0; idx--)
         {
             CCDeviceStatus* current = [_statusArray objectAtIndex:idx];
             points[idx] = [ZWL_MapUtils geoGaoDePoint2Coordinate2D:current.gaode_point];
